safari-600.7-branch/LayoutTests/media/media-source/media-source-monitor-source-buffers-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(sourceBuffer.appendBuffer(sample))
EVENT(updateend)
RUN(video.play())
EVENT(waiting) TEST(video.readyState == HTMLMediaElement.HAVE_CURRENT_DATA) OK
END OF TEST
